Move `IsModule`
