+++ crates/swc_compiler_base/src/lib.rs
@@ -1,15 +1,12 @@
 use std::{
-    env, fmt,
+    env,
     path::{Path, PathBuf},
 };
 
 use anyhow::{Context, Error};
 use base64::prelude::{Engine, BASE64_STANDARD};
 use once_cell::sync::Lazy;
-use serde::{
-    de::{Unexpected, Visitor},
-    Deserialize, Deserializer, Serialize, Serializer,
-};
+use serde::{Deserialize, Serialize};
 use swc_atoms::JsWord;
 use swc_common::{
     collections::AHashMap,
@@ -19,7 +16,8 @@ use swc_common::{
     sync::Lrc,
     BytePos, FileName, SourceFile, SourceMap,
 };
-use swc_config::{config_types::BoolOr, merge::Merge};
+use swc_config::config_types::BoolOr;
+pub use swc_config::IsModule;
 use swc_ecma_ast::{EsVersion, Ident, Program};
 use swc_ecma_codegen::{text_writer::WriteJs, Emitter, Node};
 use swc_ecma_minifier::js::JsMinifyCommentOption;
@@ -381,74 +379,6 @@ impl SourceMapsConfig {
 impl Default for SourceMapsConfig {
     fn default() -> Self {
         SourceMapsConfig::Bool(true)
-    }
-}
-
-#[derive(Clone, Debug, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum IsModule {
-    Bool(bool),
-    Unknown,
-}
-
-impl Default for IsModule {
-    fn default() -> Self {
-        IsModule::Bool(true)
-    }
-}
-
-impl Merge for IsModule {
-    fn merge(&mut self, other: Self) {
-        if *self == Default::default() {
-            *self = other;
-        }
-    }
-}
-
-impl Serialize for IsModule {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        match *self {
-            IsModule::Bool(ref b) => b.serialize(serializer),
-            IsModule::Unknown => "unknown".serialize(serializer),
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for IsModule {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct IsModuleVisitor;
-
-        impl<'de> Visitor<'de> for IsModuleVisitor {
-            type Value = IsModule;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str("a boolean or the string 'unknown'")
-            }
-
-            fn visit_bool<E>(self, b: bool) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Ok(IsModule::Bool(b))
-            }
-
-            fn visit_str<E>(self, s: &str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                match s {
-                    "unknown" => Ok(IsModule::Unknown),
-                    _ => Err(serde::de::Error::invalid_value(Unexpected::Str(s), &self)),
-                }
-            }
-        }
-
-        deserializer.deserialize_any(IsModuleVisitor)
     }
 }
